requirements/requirements.txt
# Core Telegram Libraries
pyrogram>=2.0.0
tgcrypto>=1.2.0

# Database
pymongo>=4.0.0

# AI & ML
groq>=0.3.0
openai-whisper>=20231117
pytesseract>=0.3.10
aiohttp>=3.11.11
wikipedia-api>=0.7.1
elevenlabs>=1.3.1

# Media Processing
Pillow>=8.3.0
opencv-python>=4.8.0.76
pydub>=0.25.1
gTTS>=2.5.1
gallery-dl>=1.28.2
yt-dlp[default]>=2024.12.23; python_version >= "3.8"  --pre  # Use nightly builds

# Scheduling & Async
APScheduler>=3.6.3
more-itertools>=10.2.0

# HTTP & Networking
requests>=2.26.0
urllib3>=2.0.7
beautifulsoup4>=4.12.3

# Environment & Configuration
python-dotenv>=0.19.0
schedule>=1.0.0

# Data Visualization & Scientific Computing
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0  # For color palettes and styling

# Development Tools
setuptools-rust>=1.8.0
tldextract>=5.0.0